% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmark R and Rcpp functions}
\description{
Use R package \code{microbenchmark} to compare the performance of C functions (\code{gibbsR}) and Cpp functions (\code{gibbsC}).
}
\examples{
\dontrun{
tm1 <- microbenchmark::microbenchmark(
  rnR = gibbsR(100),
  rnC = gibbsC(100)
)
print(summary(tm1)[,c(1,3,5,6)])
}
}
