% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{gibbsR}
\alias{gibbsR}
\title{A Gibbs sampler using R}
\usage{
gibbsR(N)
}
\arguments{
\item{N}{the number of samples}
}
\value{
a random sample of size \code{N}
}
\description{
A Gibbs sampler using R
}
\examples{
\dontrun{
rnR <- gibbsR(100)
par(mfrow=c(2,1));
plot(rnR[,1],type='l')
plot(rnR[,2],type='l')
}
}
