% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{Indic}
\alias{Indic}
\title{The indicator function}
\usage{
Indic(e)
}
\arguments{
\item{e}{the parameter (vector)}
}
\value{
a vector of 0 and 1
}
\description{
The indicator function 1\{e <= 0\}.
}
\examples{
\dontrun{
e<-rep(c(-1,1),c(5,5))
Indic(e)
}
}
