% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{predict_qr}
\alias{predict_qr}
\title{Predict quantiles using model averaging}
\usage{
predict_qr(x, qrma)
}
\arguments{
\item{x}{the covariate matrix (the first column should be 1)}

\item{qrma}{a list return from function \code{qrma}}
}
\value{
the prediction of the tau quantile of y given x
}
\description{
Predict quantiles using model averaging.
}
\examples{
\dontrun{
data(express)
attach(express)
y<-express$y[1:100]
y.s<-express$y[101:150]
X<-express[1:100,-1]
X<-cbind(1,X)
X.s<-express[101:150,-1]
X.s<-cbind(1,X.s)
qrma.fit<-qrma(X, y)
y.hat<-predict_qr(X.s,qrma.fit)
}
}
