% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{qrma}
\alias{qrma}
\title{Model averaging for quantile regression}
\usage{
qrma(X, y, J = 5, tau = 0.5)
}
\arguments{
\item{X}{the covariate matrix (the first column should be 1)}

\item{y}{the dependent variable vector}

\item{J}{the fold of cross-validation}

\item{tau}{the quantile}
}
\value{
A list containing: {
\code{w.JCVMA} a weight vector
\code{theta.hat} a matrix, the mth column represents the estimator of theta in the mth model
\code{y.hat} the prediction of the tau quantile of y given X
\code{PE} the in-sample prediction error
}
}
\description{
Model averaging for quantile regression by J-fold cross validation.
}
\examples{
\dontrun{
data(express)
attach(express)
y<-express$y
X<-express[,-1]
qrma.fit<-qrma(X, y)
}
}
